#874-D Flipper
https://codeforces.com/contest/1833/problem/D

You are given a permutation p of length n

.

A permutation is an array consisting of n
distinct integers from 1 to n in any order. For example, {2,3,1,5,4} is a permutation, while {1,2,2} is not (since 2 appears twice), and {1,3,4} is also not a permutation (as n=3, but the array contains 4

).

To the permutation p

, you need to apply the following operation exactly once:

    First you choose a segment [l,r]

(1≤l≤r≤n, a segment is a continuous sequence of numbers {pl,pl+1,…,pr−1,pr}) and reverse it. Reversing a segment means swapping pairs of numbers (pl,pr), (pl+1,pr−1), ..., (pl+i,pr−i) (where l+i≤r−i
).
Then you swap the prefix and suffix: [r+1,n]
and [1,l−1]

    (note that these segments may be empty). 

For example, given n=5,p={2,3,1,5,4}
, if you choose the segment [l=2,r=3], after reversing the segment p={2,1,3,5,4}, then you swap the segments [4,5] and [1,1]. Thus, p={5,4,1,3,2}

. It can be shown that this is the maximum possible result for the given permutation.

You need to output the lexicographically maximum permutation that can be obtained by applying the operation described exactly once.

A permutation a
is lexicographically greater than permutation b if there exists an i (1≤i≤n) such that aj=bj for 1≤j<i and ai>bi

.
Input

The first line of the input contains a single integer t
(1≤t≤1000

) — the number of test cases.

Then the descriptions of the test cases follow.

The first line of each test case contains a single integer n
(1≤n≤2000

) — the size of the permutation.

The second line of each test case contains n
integers: p1,p2,…,pn (1≤pi≤n) — the permutation p

itself.

It is guaranteed that the sum of n
over all test cases does not exceed 2000

.
Output

For each test case, output in a separate line the lexicographically maximum permutation of length n
that can be obtained from p

by applying the operation described in the problem exactly once.
Example
Input
Copy

9
5
2 3 1 5 4
9
4 1 6 7 2 8 5 3 9
4
4 3 2 1
2
2 1
6
3 2 4 1 5 6
7
3 2 1 5 7 6 4
10
10 2 5 6 1 9 3 8 4 7
4
4 2 1 3
1
1

Output
Copy

5 4 1 3 2 
9 4 1 6 7 2 8 5 3 
3 2 1 4 
1 2 
6 5 3 2 4 1 
7 6 4 5 3 2 1 
9 3 8 4 7 1 10 2 5 6 
3 4 2 1 
1 

Note

The first example is explained in the problem statement.

In the second example, the segment [l=9,r=9]

should be chosen.

In the third example, the segment [l=1,r=1]

should be chosen.

In the fourth example, the segment [l=1,r=2]

should be chosen.

In the fifth example, the segment [l=5,r=6]

should be chosen.

In the sixth example, the segment [l=4,r=4]

should be chosen.

In the seventh example, the segment [l=5,r=5]
should be chosen.