Educational Codeforces Round 147-C Tear It Apart
You are given a string s

, consisting of lowercase Latin letters.

In one operation, you can select several (one or more) positions in it such that no two selected positions are adjacent to each other. Then you remove the letters on the selected positions from the string. The resulting parts are concatenated without changing their order.

What is the smallest number of operations required to make all the letters in s

the same?
Input

The first line contains a single integer t
(1≤t≤104

) — the number of testcases.

The only line of each testcase contains a string s
, consisting of lowercase Latin letters. Its length is from 1 to 2⋅105

.

The total length of the strings over all testcases doesn't exceed 2⋅105

.
Output

For each testcase, print a single integer — the smallest number of operations required to make all the letters in the given string s

the same.
Example
Input
Copy

5
abacaba
codeforces
oooooooo
abcdef
mewheniseearulhiiarul

Output
Copy

1
3
0
2
4

Note

In the first testcase, you can select positions 2,4
and 6

and remove the corresponding letters 'b', 'c' and 'b'.

In the third testcase, the letters in the string are already the same, so you don't have to make any operations.

In the fourth testcase, one of the possible solutions in 2
operations is the following. You can select positions 1,4,6 first. The string becomes "bce". Then select positions 1 and 3. The string becomes "c". All letters in it are the same, since it's just one letter.