CodeTON Round 4-A Beautiful Sequence
https://codeforces.com/contest/1810/problem/A

A sequence of m integers a1,a2,…,am is good, if and only if there exists at least one i (1≤i≤m) such that ai=i. For example, [3,2,3] is a good sequence, since a2=2, a3=3, while [3,1,1] is not a good sequence, since there is no i such that ai=i

.

A sequence a
is beautiful, if and only if there exists at least one subsequence of a satisfying that this subsequence is good. For example, [4,3,2] is a beautiful sequence, since its subsequence [4,2] is good, while [5,3,4]

is not a beautiful sequence.

A sequence b
is a subsequence of a sequence a if b can be obtained from a

by the deletion of several (possibly, zero or all) elements.

Now you are given a sequence, check whether it is beautiful or not.
Input

Each test contains multiple test cases. The first line contains a single integer t
(1≤t≤500

) — the number of test cases. Their description follows.

The first line of each test case contains an integer n
(1≤n≤100

) — the length of the given sequence.

The second line of each test case contains n
integers a1,a2,…,an (1≤ai≤109

), representing the sequence.
Output

For each test case, output "YES" or "NO"(without quotes) in one line, representing whether the given sequence is beautiful.

You can output the answer in any case (upper or lower). For example, the strings "yEs", "yes", "Yes", and "YES" will be recognized as positive responses.